requirements.txt
opencv-python
d3dshot
pyturbojpeg
pyqt5
pillow
flask
waitress